Narrow integer to param class constants
You cannot just pass any integer as a type, it has to map to something
meaningful.

The psalm types introduced here are copied from doctrine/dbal 4, with a
small difference: since we are not dealing with actual enums yet, but
with class constants, we add ::*, which is compatible with both cases
(class or enum).