Restore job with --prefer-lowest
It is useful to catch misconfigured dependency constraints. It was
removed in 413c33274d02ca4d320a6ec95dff8c3951c9ca5c.
This implies configuring mocks so as to support psr/cache 1
psr/cache 1 does not use native return types, and phpdoc is not enough
to obtain a mock that has typed methods.