Original entity data resolves inverse 1-1 joins
If the source entity for an inverse (non-owning) 1-1 relationship is
identified by an association then the identifying association may not
be set when an inverse one-to-one association is resolved. This means
that no data is available in the entity to resolve the needed column
value for the join query.

Provide the original entity data for the source entity as a fallback
to resolve the query conditions.

The new test will fail if the fix is removed (comment out lines
840-847 in BasicEntityPersister.php).

Fixes #11108