Cache null/empty returns for findOneBy
On 'findOneBy' calls which return no an empty result set from the
persister, null values are not cached.
This makes it so that repeat calls of 'findOneBy' with the same
criteria, always hit the database.

The changes introduced in this commit make the AbstractEntityPersister
cache the empty result set in the region, so that repeat calls do not
hit the database again.

The existing mechanism to invalidate the query cache is sufficient; any
persist/update performed to the region will invalidate it, exactly how
it happens for non-empty result sets.

This is also more in-line with the way 'loadByCriteria' and 'loadAll'
behave.