#11977 attempting to add batching support to `BasicEntityPersister` insert operations
This change is mostly exploratory: it highlights some flaws in the API design of
`BasicEntityPersister`, showing that we need to extract the `getInsertSQL()` logic
to a new method that produces separate strings for fields and `VALUES`.

We should then be able to perform simplistic batching whenever `$isPostInsertId` is not
necessary.