#11977 implemented batching of `INSERT` statements by aggregating `VALUES()` for assigned ID entities
This change extracts `getInsertSQLFragments()` from `getInsertSQL()`, allowing us to create a larger
`INSERT INTO (columns) VALUES (values1), (values2), etc.` SQL statement.

We use a method-local anonymous class to track the current batch size, so we don't accidentally leak
the `Statement` instance, and we re-generate the statement if the batch operation fails due to the batch
being too small.

Sizing of each batch is handled by the anonymous class, mostly to avoid going
crazy with by-ref variables, or local variables that survive a single iteration.

The final result is that the `BasicEntityPersister` will execute as few statements as possible, given
the configured `BasicEntityPersister#$maxBatchSize`.

The logic is still a bit clunky, but can be refactored with some review: it is performance sensitive,
so multiple eyes should check it. Perhaps we can use a generator to make it simpler to stomach?